Refactor example in example folder as it is specific to python2 and breaks when using library with python3
