+++ ky040/KY040.py
@@ -3,7 +3,7 @@
 #stuffaboutcode.com
 
 import RPi.GPIO as GPIO
-from time import sleep
+
 
 class KY040:
 
@@ -53,28 +53,3 @@ class KY040:
 
         if GPIO.input(self.switchPin) == 0:
             self.switchCallback()
-
-#test
-if __name__ == "__main__":
-    
-    CLOCKPIN = 5
-    DATAPIN = 6
-    SWITCHPIN = 13
-
-    def rotaryChange(direction):
-        print "turned - " + str(direction)
-    def switchPressed():
-        print "button pressed"
-
-    GPIO.setmode(GPIO.BCM)
-    
-    ky040 = KY040(CLOCKPIN, DATAPIN, SWITCHPIN, rotaryChange, switchPressed)
-
-    ky040.start()
-
-    try:
-        while True:
-            sleep(0.1)
-    finally:
-        ky040.stop()
-        GPIO.cleanup()
